requirements.txt
flask==3.0.0
python-barcode==0.15.1
pillow==10.2.0
gunicorn==21.2.0